Kill mathutil
Moving only internally used function to utils.py; Rest is undocumented anyway
