+++ src/python/Makefile.am
@@ -1,6 +1,6 @@
 SUBDIRS = widget data app
 pkgpyexec_PYTHON = enumcompat.py camcalibrator.py textarea.py \
-       mathutil.py avgapp.py appstarter.py utils.py filter.py \
+       avgapp.py appstarter.py utils.py filter.py \
        mtemu.py geom.py parsecamargs.py apphelpers.py methodref.py \
        statemachine.py coordcalibrator.py graph.py __init__.py gesture.py \
        persist.py
