+++ src/python/utils.py
@@ -22,7 +22,7 @@
 
 import os
 
-from libavg import avg, mathutil, player
+from libavg import avg, player
 
 
 def getMediaDir(_file_=None, subdir='media'):
@@ -45,9 +45,36 @@ def getMediaDirFromNode(node, path=''):
     else:
         return path
 
+def getScaledDim (size, max = None, min = None):
+    width, height = size
+    if width == 0 or height == 0:
+        return size
+
+    if max:
+        max = Point2D(max)
+        assert (max.x > 0 and max.y > 0)
+        if width > max.x:
+            height = height * (max.x / width)
+            width = max.x
+        if height > max.y:
+            width = width * (max.y / height)
+            height = max.y
+
+    if min:
+        min = Point2D(min)
+        assert (min.x > 0 and min.y > 0)
+        if width < min.x:
+            height = height * (min.x / width)
+            width = min.x
+        if height < min.y:
+            width = width * (min.y / height)
+            height = min.y
+
+    return Point2D(width, height)
+
 def createImagePreviewNode(maxSize, absHref):
     node =  player.createNode('image', {'href': absHref})
-    node.size = mathutil.getScaledDim(node.size, max = maxSize)
+    node.size = getScaledDim(node.size, max = maxSize)
     return node
 
 def initFXCache(numFXNodes):
